pyproject.toml
[project]
name = "hyper-stack-manager"
version = "0.1.0"
description = "Hyper Stack Manager (hsm) - A meta-orchestrator for dynamic Python and Docker environments"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "ruamel.yaml>=0.18.0",
    "tomlkit>=0.12.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
    "tomli-w>=1.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.scripts]
hsm = "hyper_stack_manager.cli:app"

[build-system]
requires = ["uv_build>=0.9.26"]
build-backend = "uv_build"